% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squire_utils.R
\name{fit_spline_rt}
\alias{fit_spline_rt}
\title{Fit squire model using rw splines}
\usage{
fit_spline_rt(
  data,
  country,
  population,
  reporting_fraction,
  n_mcmc = 10000,
  replicates = 100,
  rw_duration = 14,
  hosp_beds = 1e+10,
  icu_beds = 1e+10
)
}
\arguments{
\item{data}{Data to fit to.  This must be constructed with
\code{particle_filter_data}}

\item{country}{Character for country beign simulated. WIll be used to
generate \code{population} and \code{contact_matrix_set} if
unprovided. Either \code{country} or \code{population} and
\code{contact_matrix_set} must be provided.}

\item{population}{Population vector (for each age group). Default = NULL,
which will cause population to be sourced from \code{country}}

\item{reporting_fraction}{Reporting fraction. Numeric for what proportion of
the total deaths the reported deaths represent. E.g. 0.5 results in
the model calibrating to twice the deaths provided by \code{data$deaths}}

\item{n_mcmc}{number of mcmc mcmc iterations to perform}

\item{replicates}{number of trajectories (replicates) to be returned that are being sampled from the posterior probability results produced by \code{run_mcmc_chain}
to select parameter set. For each parmater set sampled, run particle filter with \code{n_particles} and sample 1 trajectory}

\item{rw_duration}{Random Walk/Spline Duration. Default = 14 days}
}
\value{
Model fit from [squire:::pmcmc]
}
\description{
Fit squire model using rw splines
}
